requirements.txt
# MethodThinker Dependencies
# 安装命令: pip install -r requirements.txt
# 更新日期: 2026-04-07

# =============================================================================
# 核心依赖 (Core Dependencies)
# =============================================================================
# 必需依赖，验证系统和知识库核心功能

# PyTorch - 深度学习框架
torch>=2.1.0

# Transformers - HuggingFace 模型库
transformers>=4.40.0

# YAML 配置文件解析
pyyaml>=6.0

# 数据验证和模型定义
pydantic>=2.0.0

# 基础数据处理
numpy>=1.24.0

# =============================================================================
# 训练系统依赖 (Training Dependencies)
# =============================================================================
# 用于模型训练（SFT/方法论注入/多样性训练）

# 模型加速和分布式训练
accelerate>=0.25.0

# LoRA/QLoRA 参数高效微调
peft>=0.7.0

# TRL - Transformer Reinforcement Learning
trl>=0.7.0

# HuggingFace 数据集加载
datasets>=2.14.0

# =============================================================================
# 评估系统依赖 (Evaluation Dependencies)
# =============================================================================
# 用于模型性能评估和 Pass@K 计算

# 科学计算 - 统计分析
scipy>=1.10.0

# 符号数学 - 答案验证
sympy>=1.12

# 数据分析 - 结果处理
pandas>=2.0.0

# =============================================================================
# API客户端依赖 (API Client Dependencies)
# =============================================================================
# 用于 Layer 2 多模型验证（可选，按需安装）

# OpenAI API 客户端
openai>=1.0.0

# 阿里云 Qwen API 客户端
dashscope>=1.14.0

# =============================================================================
# 推理加速依赖 (Inference Acceleration)
# =============================================================================
# 用于 vLLM 推理加速（可选，需要 GPU）

# vLLM 高性能推理
vllm>=0.4.0

# =============================================================================
# 工具依赖 (Utility Dependencies)
# =============================================================================
# 辅助工具和进度显示

# 进度条显示
tqdm>=4.65.0

# 富文本终端输出
rich>=13.0.0

# 环境变量管理
python-dotenv>=1.0.0

# =============================================================================
# 测试依赖 (Testing Dependencies)
# =============================================================================
# 单元测试和覆盖率报告

# 测试框架
pytest>=7.0.0

# 测试覆盖率报告
pytest-cov>=4.0.0

# 异步测试支持
pytest-asyncio>=0.21.0

# =============================================================================
# 代码质量依赖 (Code Quality Dependencies)
# =============================================================================
# 类型检查、格式化、lint（可选，开发环境）

# 类型检查
mypy>=1.0.0

# 代码格式化
black>=23.0.0

# 快速 lint 工具
ruff>=0.0.260

# =============================================================================
# 可选依赖 (Optional Dependencies)
# =============================================================================
# 按特定需求安装

# TensorBoard 训练监控
# tensorboard>=2.14.0

# Weights & Biases 实验追踪
# wandb>=0.15.0

# DeepSpeed 分布式训练
# deepspeed>=0.12.0

# bitsandbytes 量化
# bitsandbytes>=0.41.0

# =============================================================================
# 安装说明
# =============================================================================
#
# 最小安装（只安装核心依赖）:
#   pip install torch>=2.1.0 transformers>=4.40.0 pyyaml>=6.0 pydantic>=2.0.0 numpy>=1.24.0
#
# 验证系统安装:
#   pip install -r requirements.txt --only-binary :all: -e ".[validation]"
#
# 训练系统安装:
#   pip install -r requirements.txt --only-binary :all: -e ".[training]"
#
# 开发环境安装（包含测试和代码质量工具）:
#   pip install -r requirements.txt
#
# GPU 训练环境:
#   pip install torch>=2.1.0 --index-url https://download.pytorch.org/whl/cu121
#   pip install -r requirements.txt
#
# =============================================================================